Add more Arduino funcs, correct CAN warning

+++ cores/arduino/api/CanMsg.h
@@ -68,7 +68,7 @@ public:
     size_t len = 0;
 
     /* Print the header. */
-    len = snprintf(buf, sizeof(buf), "[%08X] (%d) : ", id, data_length);
+    len = snprintf(buf, sizeof(buf), "[%08X] (%d) : ", (unsigned)id, data_length);
     size_t n = p.write(buf, len);
 
     /* Print the data. */
